FactFibMemo

Documentation

type Z = Integer

A convenient abbreviation.

errorNegative :: a

An error we will use repeatedly.

Factorial

fact :: Z -> Z

Write the factorial function using explicit recursion. Your function should look something like this:

fact n = ... fact ...

Examples:

>>> fact 5
120

>>> fact 10
3628800

The Y-combinator

y :: (t -> t) -> t

This is the Y-combinator. Here's the definition:

y f = f (y f)

A combinator, you will recall, is a point-free function. What is it good for? Read on!

fact' :: (Z -> Z) -> Z -> Z

Write fact', a non-recursive version of fact, that takes an additional argument as the function to call when the recursion takes place. Your code will now look like this:

fact' f n = ... f ...

Now try your function with the Y-combinator:

>>> (y fact') 6
720

Do you understand why this works? Try reducing (y fact') 3 by hand.

yesToFact :: Bool

Just to check that you do understand what's going on:

>>> yesToFact
True

Fibonacci

fib :: Int -> Z

Fibonacci is notorious, in that it takes a long time to compute. Define fib using explicit recursion:

>>> fib 6
8

Now try mapM_ (print.fib) [0,10..40]... you will wait a long time for 40! Be prepared to ^C.

fib' :: (Int -> Z) -> Int -> Z

Let's do the same transformation we did for factorial:

>>> (y fib') 6
8

Memoization

memo :: (Int -> a) -> Int -> a

Now we'll memoize this function. I'm going to give you the definition. If you copy this into the code itself, the example will work. The question is, do you understand why? It's not obvious. Work your way through it. I suggest reducing a small example by hand:

(y (memo . fib')) 3

Here's the definition:

memo f = (map f [0..] !!)

>>> (y (memo . fib')) 6
8

>>> (y (memo . fib')) 100
354224848179261915075

yesToMemo :: Bool

Just to check that you do understand what's going on:

>>> yesToMemo
True